@@ -15,7 +15,7 @@
     "test:ts": "jest",
     "test": "npm run test:ts",
     "lint": "npm run lint:ts && npm run lint:scss",
-    "lint:fix-ts": "gts fix",
+    "lint:ts-fix": "gts fix",
     "lint:ts": "gts check",
     "lint:scss": "stylelint src/**/*.scss",
     "prepare": "echo No prepare scripts set",
@@ -39,10 +39,9 @@
     "parcel-bundler": "^1.10.0",
     "postcss-modules": "^1.4.1",
     "posthtml-favicons": "^1.3.0",
-    "react": "^16.4.0",
-    "react-dom": "^15.3.0 || 16",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^5.2.1",
-    "react-router-dom": "^5.1.2",
     "react-schemaorg": "^1.0.0",
     "sass": "^1.23.0",
     "schema-dts": "^0.4.2",
@@ -52,23 +51,26 @@
     "uuid": "^8.3.1"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@testing-library/jest-dom": "^4.1.0",
-    "@testing-library/react": "^9.1.3",
-    "@testing-library/react-hooks": "^2.0.1",
-    "@types/classnames": "^2.2.9",
-    "@types/jest": "^24.9.1",
-    "@types/jsdom": "^12.2.4",
-    "@types/node": "^10.0.3",
-    "@types/react": "^16.9.9",
-    "@types/react-dom": "^16.9.2",
-    "@types/react-helmet": "^5.0.9",
-    "@types/react-router-dom": "^5.1.0",
-    "@typescript-eslint/eslint-plugin": "^2.0.0",
-    "@typescript-eslint/parser": "^2.0.0",
-    "babel-jest": "^24.9.0",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@types/classnames": "^2.3.1",
+    "@types/jest": "^28.1.6",
+    "@types/jsdom": "^20.0.0",
+    "@types/node": "^18.6.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-helmet": "^6.1.5",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/uuid": "^8.3.4",
+    "@typescript-eslint/eslint-plugin": "^5.32.0",
+    "@typescript-eslint/parser": "^5.32.0",
+    "babel-jest": "^28.1.3",
     "eslint-config-prettier": "^6.1.0",
-    "gts": "^1.1.2",
+    "eslint-plugin-jest": "^26.7.0",
+    "gts": "^3.1.0",
+    "history": "^5.3.0",
     "husky": "^3.0.4",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^24.9.0",
@@ -76,7 +78,7 @@
     "jest-css-modules-transform": "^2.5.0",
     "jest-date-mock": "^1.0.7",
     "jest-extended": "^0.11.2",
-    "jest-fetch-mock": "^2.1.2",
+    "jest-fetch-mock": "^3.0.3",
     "jest-svg-transformer": "^1.0.0",
     "jest-transform-css": "^2.0.0",
     "jest-transform-stub": "^2.0.0",
@@ -84,8 +86,10 @@
     "postcss-modules": "^1.4.1",
     "prettier": "^1.18.2",
     "react-element-to-jsx-string": "^14.0.3",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-test-renderer": "^16.9.0",
-    "stylelint": "^10.1.0",
+    "stylelint": "^13.7.2",
     "stylelint-config-recommended-scss": "^3.3.0",
     "stylelint-scss": "^3.9.4"
   }
