@@ -1,6 +1,6 @@
 module.exports = {
   hooks: {
-    'pre-commit': 'npm run lint:fix-ts',
+    'pre-commit': 'npm run lint:ts-fix',
     'pre-push': 'npm test',
   },
 };
