@@ -1,4 +1,3 @@
-const fs = require('fs');
 const path = require('path');
 const srcPath = path.resolve(__dirname, 'src');
 
@@ -6,7 +5,7 @@ module.exports = {
   sassConfig: {
     includePaths: [srcPath],
     importer: [
-      (childResource, parentResource) => {
+      childResource => {
         const file = childResource.replace(/^~/, srcPath);
 
         return { file };
