@@ -1,17 +1,13 @@
 import React, { ReactElement } from 'react';
-import { BrowserRouter, RouteComponentProps } from 'react-router-dom';
+import { BrowserRouter, PathRouteProps } from 'react-router-dom';
 import { renderToString } from 'react-dom/server';
-import {
-  createMemoryHistory,
-  MemoryHistoryBuildOptions,
-  Location,
-} from 'history';
-import { match } from 'react-router';
+import { createMemoryHistory, Location, MemoryHistoryOptions } from 'history';
+// import { match } from 'react-router-dom';
 import { RenderResult } from '@testing-library/react';
-// tslint:disable-next-line:ban-ts-ignore
+// eslint-disable-next-line
 // @ts-ignore
 import reactElementToJSXString from 'react-element-to-jsx-string';
-import { Optional, POJSObject } from '~/app/declarations/standard';
+import { Optional } from '~/app/declarations/standard';
 
 export interface RouteWrapperConfig {
   children?: JSX.Element[];
@@ -36,7 +32,10 @@ export function getElementByQuery(
   return el.querySelector(query);
 }
 
-function mutateProps(props: POJSObject, stringifyObjects = true): void {
+function mutateProps(
+  props: Record<string, any>,
+  stringifyObjects = true
+): void {
   let key: string;
   // tslint:disable-next-line:forin
   for (key in props) {
@@ -61,15 +60,15 @@ function mutateProps(props: POJSObject, stringifyObjects = true): void {
 
       props[key] = parsedVal;
     } else if (stringifyObjects && typeof val === 'object') {
-      props[key] = JSON.stringify(val as {});
+      props[key] = JSON.stringify(val as Record<string, unknown>);
     } else {
       props[key] = val;
     }
   }
 }
 
-export interface MockRouteComponentProps<P> {
-  history?: Partial<MemoryHistoryBuildOptions>;
+export interface MockRouteComponentProps<P extends string> {
+  history?: Partial<MemoryHistoryOptions>;
   location?: Partial<Location>;
   match?: Partial<match<P>>;
 }
@@ -78,7 +77,7 @@ export function createRouteComponentProps<T>({
   history,
   location,
   match,
-}: MockRouteComponentProps<T> = {}): RouteComponentProps<T> {
+}: MockRouteComponentProps<T> = {}): PathRouteProps {
   const defaultLocation = {
     pathname: '',
     search: '',
@@ -115,8 +114,8 @@ export function createMockComponent(
 export function createMockContextConsumer(
   componentName: string,
   testId: string,
-  mockState: {} = {}
-): jest.Mock<React.Consumer<{}>> {
+  mockState: Record<string, unknown> = {}
+): jest.Mocked<React.Consumer<Record<string, unknown>>> {
   return jest.fn(props => {
     const { children } = props;
 
@@ -126,21 +125,19 @@ export function createMockContextConsumer(
   });
 }
 
-export function getMockCallValue<T extends unknown>(
-  mock: jest.Mock,
-  index = 0
-): T[] {
+export function getMockCallValue<T>(mock: jest.Mocked, index = 0): T[] {
   return mock.mock.calls[index] as T[];
 }
 
-export function getMockReturnValue<T extends unknown>(
-  mock: jest.Mock,
-  index = 0
-): T {
+export function getMockReturnValue<T>(mock: jest.Mocked, index = 0): T {
   return mock.mock.results[index].value as T;
 }
 
-export function createMockAppEnums() {
+export function createMockAppEnums(): {
+  BaseReducerActionType: Record<string, string>;
+  AsyncReducerActionType: Record<string, string>;
+  EditableReducerActionType: Record<string, string>;
+} {
   enum mockBaseReducerActionType {
     Create = 'mock-create',
     Delete = 'mock-delete',
@@ -151,7 +148,6 @@ export function createMockAppEnums() {
   enum mockAsyncReducerActionType {
     Request = 'mock-request',
     Receive = 'mock-receive',
-    Request = 'mock-request',
   }
 
   enum mockEditableReducerActionType {
@@ -159,11 +155,9 @@ export function createMockAppEnums() {
     Reset = 'mock-reset',
   }
 
-  const appModule = {
+  return {
     BaseReducerActionType: mockBaseReducerActionType,
     AsyncReducerActionType: mockAsyncReducerActionType,
     EditableReducerActionType: mockEditableReducerActionType,
   };
-
-  return appModule;
 }
