@@ -1,3 +1,3 @@
-import mockFetch from "jest-fetch-mock";
+import mockFetch from 'jest-fetch-mock';
 
 global.fetch = mockFetch;
