@@ -1,5 +1,5 @@
-import { RouteComponentProps, Switch } from "react-router";
-import { FunctionComponent, ReactElement } from "react";
+import { RouteComponentProps, Switch } from 'react-router';
+import { FunctionComponent, ReactElement } from 'react';
 
 export interface AppHistoryProps extends RouteComponentProps {
   children: ReactElement | ReactElement[];
