@@ -1,19 +1,30 @@
-import { RouteComponentProps, RouteProps, Switch } from 'react-router';
+import {
+  PathRouteProps,
+  LayoutRouteProps,
+  IndexRouteProps,
+  RouteProps,
+  Switch,
+} from 'react-router-dom';
 import { FunctionComponent, default as React, ReactElement } from 'react';
-import { Optional } from '~/app/declarations/common';
 
 export type AnyView =
   // tslint:disable-next-line:no-any
-  | React.ComponentType<RouteComponentProps<any>>
+  | React.ComponentType<ReactRouterRoutesUnion<any>>
   // tslint:disable-next-line:no-any
   | React.ComponentType<any>;
 
-export interface ExtendedRouteProps extends RouteProps {
+type ReactRouterRoutesUnion =
+  | PathRouteProps
+  | LayoutRouteProps
+  | IndexRouteProps;
+
+export interface ExtendedRouteProps extends PathRouteProps {
   readonly name?: string;
   readonly matchingValues?: string;
   readonly requiresAuth?: boolean;
   readonly scrollToTop?: boolean;
   readonly disabled?: boolean;
+  readonly component: AnyView;
 }
 
 interface UrlPatternOptions {
@@ -27,6 +38,7 @@ interface UrlPatternOptions {
 }
 
 export interface UrlPattern {
+  // eslint-disable-next-line @typescript-eslint/no-misused-new
   new (path: string, options?: UrlPatternOptions): UrlPattern;
-  match(path: string): Optional<{}>;
+  match(path: string): Record<string, unknown> | undefined;
 }
