@@ -3,7 +3,7 @@ import { History } from 'history';
 import { useAuth } from '~/app/stores/auth/AuthContext';
 import { AuthContextValue } from '~/app/stores/auth/AuthContext.d';
 import { UseRedirectToAuthHookResult } from './AppRoutesHooks.d';
-import { redirectToAuth } from '~/app/shared/utils/Auth';
+import { redirectToAuth, RedirectLoginOptions } from '~/app/shared/utils/Auth';
 
 function createRedirectToAuth(
   { location }: History,
