@@ -5,6 +5,7 @@ import { useUserData } from '~/app/stores/user/UserContext';
 import { AnyView, ExtendedRouteProps } from './AppRoutes.d';
 import { redirectToAuth } from '~/app/shared/utils/Auth';
 import { OptionalN } from '~/app/declarations/standard';
+import { useLocation } from 'react-router-dom';
 
 export function AuthenticatedRoute({ component, ...rest }: ExtendedRouteProps) {
   const View: AnyView = component as AnyView;
@@ -18,10 +19,12 @@ export function AuthenticatedRoute({ component, ...rest }: ExtendedRouteProps) {
   let view: OptionalN<JSX.Element>;
 
   if (isReady && !isAuthenticated && !loading) {
-    const { location } = rest;
+    const location = useLocation();
+
     redirectToAuth(location, loginWithRedirect);
     view = null;
   } else if (isAuthenticated && user) {
+    //view = <Route {...rest} element={<View />} />;
     view = <View {...rest} />;
   } else {
     view = null;
