@@ -1,2 +1,2 @@
-declare module '*.svg'
-declare module '*.png'
+declare module '*.svg';
+declare module '*.png';
