@@ -1,5 +1,5 @@
 declare namespace jest {
   interface Matchers<R> {
-    toEqualJSX(expected: React.ReactElement): R
+    toEqualJSX(expected: React.ReactElement): R;
   }
 }
