@@ -4,7 +4,7 @@ import {
   FetchPromise,
   FetchCoreArgs,
 } from '~/app/shared/services/fetch/Fetch.d';
-import { JSONData, OptionalU } from '~/app/declarations/standard';
+import { JSONData } from '~/app/declarations/standard';
 
 export interface APIFetchOptions {
   requiresAuth?: boolean;
@@ -36,8 +36,11 @@ export interface APIStructuredResponseErrorItem {
   field_name?: string;
 }
 
-export interface APIStructuredResponse<R = unknown, M = APIStructuredResponseMetaDefault> {
-  meta: M,
-  result?: R,
-  errors?: APIStructuredResponseErrorItem[],
+export interface APIStructuredResponse<
+  R = unknown,
+  M = APIStructuredResponseMetaDefault
+> {
+  meta: M;
+  result?: R;
+  errors?: APIStructuredResponseErrorItem[];
 }
