@@ -1,6 +1,7 @@
 import React from 'react';
 import '@auth/auth-spa-js';
 import { OptionalN } from '~/app/declarations/standard';
+import { RedirectLoginOptions } from '~/app/shared/utils/Auth';
 
 interface AuthData {
   isReady: boolean;
@@ -8,6 +9,7 @@ interface AuthData {
   isAuthError: boolean;
   user: unknown;
   loading: boolean;
+  loginWithRedirect?: (options?: RedirectLoginOptions) => void;
 }
 
 export interface AuthRedirectAppState<S = unknown> {
@@ -19,9 +21,7 @@ interface AuthInternalData extends AuthData {
   appState?: AuthRedirectAppState;
 }
 
-interface AuthContextValue extends AuthData {
-  //TODO
-}
+type AuthContextValue = AuthData;
 
 type AuthContext = React.Context<OptionalN<AuthContextValue>>;
 
