@@ -1,5 +1,6 @@
-import { FC } from "react";
+import { FC } from 'react';
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface HomeFCProps {}
 
 export type ServerErrorFC = FC<HomeFCProps>;
