@@ -1,6 +1,8 @@
-import { FC } from "react";
-import { RouteComponentProps } from "react-router";
+import { FC } from 'react';
+// eslint-disable-next-line node/no-extraneous-import
+import { RouteComponentProps } from 'react-router';
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface ServerErrorFCProps {}
 
 export type ServerErrorFC = FC<ServerErrorFCProps>;
